pyproject.toml
[project]
name = "story-teller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[extensions]>=1.12.0",
    "openai>=1.99.0",
    "pillow>=12.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
